requirements.txt
setuptools==20.10.1
Django==1.10.4
djangorestframework==3.5.3
